Create Question 1
Write a program that prints out all the numbers that are multiples of 6 from the list in the same order as they appear in the list; stop printing after 949 in the sequence.
* List always contain 949.
* 949 should also be printed as the last number.
Sample list: numbers = [ 386, 562, 47, 498, 807, 344, 230, 377, 754, 484, 597, 876, 353, 615, 953, 345, 399, 162, 758, 219, 918, 237, 412, 566, 826, 248, 866, 950, 626, 949, 379, 843, 831, 445]
Sample output: 498 876 162 918 949
Note: The given list above is one of the sample lists. Your logic should work for any list that holds different numbers than above.