Create Question 9
Alex is standing at the ATM to withdraw some cash. ATM will only accept the transaction if the amount Alex is trying to withdraw is a multiple of $5, and his account balance has enough cash for the transaction (including the transaction fee).
If there is not enough cash balance in the account to complete the transaction, it does not perform the withdrawal transaction.
For each successful withdrawal, there is a transaction fee of $1.
Write a program that takes two integers from user(amount Alex wishes to withdraw, and the initial account balance) and prints out Alex’s account balance after an attempted transaction. 
* Account balance NEVER goes to negative value.
* Withdrawal amount of $0 is not considered in this question.

Sample input:         Sample output:
30 120                          89
42 120                         120
500 120                       120