Create Question 12
Write a program that takes one string from user as input and prints out the alphabet that is used the maximum number of times in that string (string can have mixed upper and lower cases). Your output alphabet should be in uppercase. If there are more than one alphabet that is used the most in the word, print '?' 

Sample input:            Sample output:
Mississippi                      ?
zZa                                  Z
z                                      Z
baaa                                A